chore: refactor: getTimelineForGroup was referenced client-side, therefore move it into /lib
